Second refactor of tests
