+++ test/integration/fda.integration.shared.js
@@ -41,7 +41,6 @@ import pg from 'pg';
 import { MongoClient } from 'mongodb';
 import http from 'node:http';
 import { spawn } from 'node:child_process';
-import net from 'node:net';
 import path from 'node:path';
 import { registerPlatformIntegrationTests } from './suites/platform.integration.tests.js';
 import { registerSlidingWindowsIntegrationTests } from './suites/slidingWindows.integration.tests.js';
@@ -344,7 +343,41 @@ export function runFDAIntegrationSuite({ mode, label }) {
       buildDaDataUrl,
       getPgHost: () => pgHost,
       getPgPort: () => pgPort,
-      connectWithRetry,
+    });
+
+    registerDefaultDataAccessIntegrationTests({
+      getBaseUrl: () => baseUrl,
+      service,
+      servicePath,
+      visibility,
+      httpReq,
+      waitUntilFDACompleted,
+      buildDaDataUrl,
+    });
+
+    registerDaDataQueriesIntegrationTests({
+      getBaseUrl: () => baseUrl,
+      service,
+      servicePath,
+      visibility,
+      fdaId,
+      daId,
+      httpReq,
+      httpReqRaw,
+      waitUntilFDACompleted,
+      buildDaDataUrl,
+    });
+
+    registerDaParamsIntegrationTests({
+      getBaseUrl: () => baseUrl,
+      service,
+      servicePath,
+      visibility,
+      fdaId,
+      daId2,
+      httpReq,
+      waitUntilFDACompleted,
+      buildDaDataUrl,
     });
 
     test('POST /fdas?defaultDataAccess=false creates FDA without default DA', async () => {
@@ -448,259 +481,6 @@ export function runFDAIntegrationSuite({ mode, label }) {
         await httpReq({
           method: 'DELETE',
           url: `${baseUrl}/${visibility}/fdas/${onlyFreshFdaId}`,
-          headers: {
-            'Fiware-Service': service,
-            'Fiware-ServicePath': servicePath,
-          },
-        });
-      }
-    });
-
-    test('defaultDataAccess without timeColumn includes pageSize/pageStart but not start/finish', async () => {
-      const untimedFdaId = 'fda_default_da_untimed';
-
-      try {
-        const createFda = await httpReq({
-          method: 'POST',
-          url: `${baseUrl}/${visibility}/fdas`,
-          headers: {
-            'Fiware-Service': service,
-            'Fiware-ServicePath': servicePath,
-          },
-          body: {
-            id: untimedFdaId,
-            query:
-              'SELECT id, name, age, authorized FROM public.users ORDER BY id',
-            description: 'default DA untimed test',
-          },
-        });
-
-        expect(createFda.status).toBe(202);
-        const completedFDA = await waitUntilFDACompleted({
-          baseUrl,
-          service,
-          fdaId: untimedFdaId,
-        });
-
-        const defaultDa = completedFDA?.das?.defaultDataAccess;
-        expect(defaultDa).toBeDefined();
-        expect(defaultDa.query).toContain('COUNT(*) OVER() as __total');
-        expect(defaultDa.query).toContain(
-          'LIMIT CAST($pageSize AS BIGINT) OFFSET CAST($pageStart AS BIGINT)',
-        );
-        expect(defaultDa.query).not.toContain('$start');
-        expect(defaultDa.query).not.toContain('$finish');
-        expect(defaultDa.params.some((p) => p.name === 'pageSize')).toBe(true);
-        expect(defaultDa.params.some((p) => p.name === 'pageStart')).toBe(true);
-        expect(defaultDa.params.some((p) => p.name === 'start')).toBe(false);
-        expect(defaultDa.params.some((p) => p.name === 'finish')).toBe(false);
-
-        const pagingRes = await httpReq({
-          method: 'GET',
-          url: buildDaDataUrl(
-            baseUrl,
-            servicePath,
-            untimedFdaId,
-            'defaultDataAccess',
-            { pageSize: 1, pageStart: 2 },
-          ),
-          headers: { 'Fiware-Service': service },
-        });
-
-        expect(pagingRes.status).toBe(200);
-        expect(pagingRes.json).toHaveLength(1);
-        expect(pagingRes.json[0].id).toBe('3');
-      } finally {
-        await httpReq({
-          method: 'DELETE',
-          url: `${baseUrl}/${visibility}/fdas/${untimedFdaId}`,
-          headers: {
-            'Fiware-Service': service,
-            'Fiware-ServicePath': servicePath,
-          },
-        });
-      }
-    });
-
-    test('defaultDataAccess includes start/finish and pageSize/pageStart when FDA has timeColumn', async () => {
-      const timedFdaId = 'fda_default_da_timed';
-
-      try {
-        const createFda = await httpReq({
-          method: 'POST',
-          url: `${baseUrl}/${visibility}/fdas`,
-          headers: {
-            'Fiware-Service': service,
-            'Fiware-ServicePath': servicePath,
-          },
-          body: {
-            id: timedFdaId,
-            query:
-              'SELECT id, name, age, timeinstant, authorized FROM public.users ORDER BY id',
-            description: 'default DA timed test',
-            timeColumn: 'timeinstant',
-          },
-        });
-
-        expect(createFda.status).toBe(202);
-        const completedFDA = await waitUntilFDACompleted({
-          baseUrl,
-          service,
-          fdaId: timedFdaId,
-        });
-
-        const defaultDa = completedFDA?.das?.defaultDataAccess;
-        expect(defaultDa).toBeDefined();
-        expect(defaultDa.query).toContain(
-          '($start IS NULL OR CAST("timeinstant" AS TIMESTAMP) >= CAST($start AS TIMESTAMP))',
-        );
-        expect(defaultDa.query).toContain(
-          '($finish IS NULL OR CAST("timeinstant" AS TIMESTAMP) <= CAST($finish AS TIMESTAMP))',
-        );
-        expect(defaultDa.query).toContain('COUNT(*) OVER() as __total');
-        expect(defaultDa.query).toContain(
-          'LIMIT CAST($pageSize AS BIGINT) OFFSET CAST($pageStart AS BIGINT)',
-        );
-        expect(defaultDa.params.some((p) => p.name === 'start')).toBe(true);
-        expect(defaultDa.params.some((p) => p.name === 'finish')).toBe(true);
-        expect(defaultDa.params.some((p) => p.name === 'pageSize')).toBe(true);
-        expect(defaultDa.params.some((p) => p.name === 'pageStart')).toBe(true);
-      } finally {
-        await httpReq({
-          method: 'DELETE',
-          url: `${baseUrl}/${visibility}/fdas/${timedFdaId}`,
-          headers: {
-            'Fiware-Service': service,
-            'Fiware-ServicePath': servicePath,
-          },
-        });
-      }
-    });
-
-    test('defaultDataAccess supports start/finish and pageSize/pageStart in cached mode', async () => {
-      const timedFdaId = 'fda_default_da_timed_data';
-
-      try {
-        const createFda = await httpReq({
-          method: 'POST',
-          url: `${baseUrl}/${visibility}/fdas`,
-          headers: {
-            'Fiware-Service': service,
-            'Fiware-ServicePath': servicePath,
-          },
-          body: {
-            id: timedFdaId,
-            query:
-              'SELECT id, name, age, timeinstant, authorized FROM public.users ORDER BY id',
-            description: 'default DA timed data test',
-            timeColumn: 'timeinstant',
-          },
-        });
-
-        expect(createFda.status).toBe(202);
-        await waitUntilFDACompleted({
-          baseUrl,
-          service,
-          fdaId: timedFdaId,
-        });
-
-        const rangeAndPagingRes = await httpReq({
-          method: 'GET',
-          url: buildDaDataUrl(
-            baseUrl,
-            servicePath,
-            timedFdaId,
-            'defaultDataAccess',
-            {
-              start: '2020-01-01T00:00:00.000Z',
-              finish: '2020-12-31T23:59:59.000Z',
-              pageSize: 1,
-              pageStart: 1,
-            },
-          ),
-          headers: { 'Fiware-Service': service },
-        });
-
-        if (rangeAndPagingRes.status >= 400) {
-          console.error(
-            'GET defaultDataAccess with start/finish/pageSize/pageStart failed:',
-            rangeAndPagingRes.status,
-            rangeAndPagingRes.json ?? rangeAndPagingRes.text,
-          );
-        }
-
-        expect(rangeAndPagingRes.status).toBe(200);
-        expect(Array.isArray(rangeAndPagingRes.json)).toBe(true);
-        expect(rangeAndPagingRes.json).toHaveLength(1);
-        expect(rangeAndPagingRes.json[0].id).toBe('2');
-      } finally {
-        await httpReq({
-          method: 'DELETE',
-          url: `${baseUrl}/${visibility}/fdas/${timedFdaId}`,
-          headers: {
-            'Fiware-Service': service,
-            'Fiware-ServicePath': servicePath,
-          },
-        });
-      }
-    });
-
-    test('defaultDataAccess matches exact equality on declared timeColumn in cached mode', async () => {
-      const timedFdaId = 'fda_default_da_timed_equality';
-
-      try {
-        const createFda = await httpReq({
-          method: 'POST',
-          url: `${baseUrl}/${visibility}/fdas`,
-          headers: {
-            'Fiware-Service': service,
-            'Fiware-ServicePath': servicePath,
-          },
-          body: {
-            id: timedFdaId,
-            query:
-              'SELECT id, name, age, timeinstant, authorized FROM public.users ORDER BY id',
-            description: 'default DA timed equality test',
-            timeColumn: 'timeinstant',
-          },
-        });
-
-        expect(createFda.status).toBe(202);
-        await waitUntilFDACompleted({
-          baseUrl,
-          service,
-          fdaId: timedFdaId,
-        });
-
-        const equalityRes = await httpReq({
-          method: 'GET',
-          url: buildDaDataUrl(
-            baseUrl,
-            servicePath,
-            timedFdaId,
-            'defaultDataAccess',
-            {
-              timeinstant: '2020-08-17T18:25:28.332Z',
-            },
-          ),
-          headers: { 'Fiware-Service': service },
-        });
-
-        if (equalityRes.status >= 400) {
-          console.error(
-            'GET defaultDataAccess with exact timeColumn equality failed:',
-            equalityRes.status,
-            equalityRes.json ?? equalityRes.text,
-          );
-        }
-
-        expect(equalityRes.status).toBe(200);
-        expect(Array.isArray(equalityRes.json)).toBe(true);
-        expect(equalityRes.json.map((row) => row.id)).toEqual(['1', '2', '3']);
-      } finally {
-        await httpReq({
-          method: 'DELETE',
-          url: `${baseUrl}/${visibility}/fdas/${timedFdaId}`,
           headers: {
             'Fiware-Service': service,
             'Fiware-ServicePath': servicePath,
@@ -917,95 +697,6 @@ export function runFDAIntegrationSuite({ mode, label }) {
       expect(secondQuery.json).toEqual([{ id: '1', name: 'ana', age: '30' }]);
     });
 
-    test('PUT /fdas/:fdaId/das/:daId updates DA with incorrect params', async () => {
-      const daIdToUpdate = 'da_update';
-
-      // Update DA (bad range param)
-      const rangeUpdateRes = await httpReq({
-        method: 'PUT',
-        url: `${baseUrl}/${visibility}/fdas/${fdaId}/das/${daIdToUpdate}`,
-        headers: {
-          'Content-Type': 'application/json',
-          'Fiware-Service': service,
-        },
-        body: {
-          id: 'ignored_in_put',
-          description: 'updated filter',
-          params: [
-            {
-              name: 'activity',
-              type: 'Number',
-              range: ['badRange', 14],
-            },
-          ],
-        },
-      });
-
-      if (rangeUpdateRes.status >= 400) {
-        console.error(
-          'PUT /das failed as expected:',
-          rangeUpdateRes.status,
-          rangeUpdateRes.json ?? rangeUpdateRes.text,
-        );
-      }
-
-      expect(rangeUpdateRes.status).toBe(400);
-
-      // Update DA (bad range enum)
-      const enumUpdateRes = await httpReq({
-        method: 'PUT',
-        url: `${baseUrl}/${visibility}/fdas/${fdaId}/das/${daIdToUpdate}`,
-        headers: {
-          'Content-Type': 'application/json',
-          'Fiware-Service': service,
-        },
-        body: {
-          id: 'ignored_in_put',
-          description: 'updated filter',
-          params: [
-            {
-              name: 'activity',
-              type: 'Number',
-              enum: ['option', true],
-            },
-          ],
-        },
-      });
-
-      if (enumUpdateRes.status >= 400) {
-        console.error(
-          'PUT /das failed as expected:',
-          enumUpdateRes.status,
-          enumUpdateRes.json ?? enumUpdateRes.text,
-        );
-      }
-
-      expect(enumUpdateRes.status).toBe(400);
-    });
-
-    test('GET /{visibility}/fdas/{fdaId}/das/{daId}/data returns JSON array when Accept: application/json', async () => {
-      const res = await httpReq({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaId, daId, {
-          minAge: 25,
-        }),
-        headers: { 'Fiware-Service': service, Accept: 'application/json' },
-      });
-
-      if (res.status >= 400) {
-        console.error(
-          'GET /{visibility}/fdas/{fdaId}/das/{daId}/data (json) failed:',
-          res.status,
-          res.json ?? res.text,
-        );
-      }
-      expect(res.status).toBe(200);
-      expect(res.json).toEqual([
-        { id: '1', name: 'ana', age: '30' },
-        { id: '3', name: 'carlos', age: '40' },
-      ]);
-    });
-
     test('GET /{visibility}/fdas/{fdaId}/data runs FDA query directly against PostgreSQL source', async () => {
       const daFreshId = 'da_fresh_users';
       const freshFdaId = 'fda_fresh_users_direct';
@@ -1136,94 +827,6 @@ export function runFDAIntegrationSuite({ mode, label }) {
 
       expect(freshRes.status).toBe(400);
       expect(freshRes.json.error).toBe('BadRequest');
-    });
-
-    test('POST /fdas/:fdaId/das rejects incompatible default values', async () => {
-      const fdaBadDefaultId = 'fda_bad_default';
-
-      const createFda = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas`,
-        headers: {
-          'Fiware-Service': service,
-          'Fiware-ServicePath': servicePath,
-        },
-        body: {
-          id: fdaBadDefaultId,
-          description: 'invalid default test fda',
-          query:
-            'SELECT id, name, age, timeinstant, authorized FROM public.users ORDER BY id',
-        },
-      });
-
-      expect(createFda.status).toBe(202);
-      await waitUntilFDACompleted({
-        baseUrl,
-        service,
-        fdaId: fdaBadDefaultId,
-      });
-
-      const createDa = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas/${fdaBadDefaultId}/das`,
-        headers: { 'Fiware-Service': service },
-        body: {
-          id: 'da_bad_default',
-          description: 'should reject invalid default',
-          query: `
-          SELECT id, name
-          WHERE authorized = $authorized
-          ORDER BY id
-        `,
-          params: [
-            {
-              name: 'authorized',
-              type: 'Boolean',
-              default: 'notBool',
-            },
-          ],
-        },
-      });
-
-      expect(createDa.status).toBe(400);
-      expect(createDa.json.error).toBe('InvalidParam');
-      expect(createDa.json.description).toContain(
-        'Default value for param "authorized" not of valid type (Boolean).',
-      );
-    });
-
-    test('POST /fdas/:fdaId/das rejects empty body', async () => {
-      const fdaBadDefaultId = 'fda_bad_body';
-
-      const createFda = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas`,
-        headers: {
-          'Fiware-Service': service,
-          'Fiware-ServicePath': servicePath,
-        },
-        body: {
-          id: fdaBadDefaultId,
-          description: 'invalid default test fda',
-          query:
-            'SELECT id, name, age, timeinstant, authorized FROM public.users ORDER BY id',
-        },
-      });
-
-      expect(createFda.status).toBe(202);
-      await waitUntilFDACompleted({
-        baseUrl,
-        service,
-        fdaId: fdaBadDefaultId,
-      });
-
-      const createDa = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas/${fdaBadDefaultId}/das`,
-        headers: { 'Fiware-Service': service },
-      });
-
-      expect(createDa.status).toBe(400);
     });
 
     test('GET /{visibility}/fdas/{fdaId}/data returns 429 when max concurrent fresh queries is reached', async () => {
@@ -1612,702 +1215,6 @@ export function runFDAIntegrationSuite({ mode, label }) {
         await pgClient.query(`DROP TABLE IF EXISTS public.${fixtureTable}`);
         await pgClient.end();
       }
-    });
-
-    test('GET /{visibility}/fdas/{fdaId}/das/{daId}/data rejects fresh query param', async () => {
-      const res = await httpReq({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaId, daId, {
-          minAge: 25,
-          fresh: 'true',
-        }),
-        headers: { 'Fiware-Service': service },
-      });
-
-      expect(res.status).toBe(400);
-      expect(res.json.error).toBe('BadRequest');
-    });
-
-    test('POST /fdas/:fdaId/das + GET /{visibility}/fdas/{fdaId}/das/{daId}/data using default params', async () => {
-      // DuckDB reads parquet generated in  s3://<bucket>/<fdaID>.parquet
-      const daQuery = `
-      SELECT id, name, age
-      WHERE age > $minAge AND name = $name AND timeinstant = $timeinstant AND authorized = $authorized
-      ORDER BY id;
-    `;
-
-      const createDa = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas/${fdaId}/das`,
-        headers: { 'Fiware-Service': service },
-        body: {
-          id: daId2,
-          description: 'get user',
-          query: daQuery,
-          params: [
-            {
-              name: 'name',
-              type: 'Text',
-              required: true,
-              enum: ['ana', 'carlos'],
-            },
-            {
-              name: 'minAge',
-              type: 'Number',
-              default: 25,
-              range: [20, 50],
-            },
-            {
-              name: 'timeinstant',
-              type: 'DateTime',
-              default: '2020-08-17T18:25:28.332Z',
-            },
-            {
-              name: 'authorized',
-              type: 'Boolean',
-              default: true,
-            },
-          ],
-        },
-      });
-
-      if (createDa.status >= 400) {
-        console.error(
-          'POST /das failed:',
-          createDa.status,
-          createDa.json ?? createDa.text,
-        );
-      }
-      expect(createDa.status).toBe(201);
-
-      // Create DA with params but no Type
-      const noTypeDa = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas/${fdaId}/das`,
-        headers: { 'Fiware-Service': service },
-        body: {
-          id: `${daId2}_noType`,
-          description: 'get user',
-          query: daQuery,
-          params: [
-            {
-              name: 'minAge',
-              default: 25,
-              range: ['badRange', 50],
-            },
-          ],
-        },
-      });
-
-      if (noTypeDa.status >= 400) {
-        console.error(
-          'POST /das failed as expected:',
-          noTypeDa.status,
-          noTypeDa.json ?? noTypeDa.text,
-        );
-      }
-      expect(noTypeDa.status).toBe(400);
-
-      // Create DA with bad params Type (invalid value)
-      const badTypeDa = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas/${fdaId}/das`,
-        headers: { 'Fiware-Service': service },
-        body: {
-          id: `${daId2}_badType`,
-          description: 'get user',
-          query: daQuery,
-          params: [
-            {
-              name: 'minAge',
-              type: 'FakeType',
-              default: 25,
-              range: ['badRange', 50],
-            },
-          ],
-        },
-      });
-
-      if (badTypeDa.status >= 400) {
-        console.error(
-          'POST /das failed as expected:',
-          badTypeDa.status,
-          badTypeDa.json ?? badTypeDa.text,
-        );
-      }
-      expect(badTypeDa.status).toBe(400);
-
-      // Create DA with incompatible default value for declared type
-      const badDefaultDa = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas/${fdaId}/das`,
-        headers: { 'Fiware-Service': service },
-        body: {
-          id: `${daId2}_badDefault`,
-          description: 'get user',
-          query: daQuery,
-          params: [
-            {
-              name: 'authorized',
-              type: 'Boolean',
-              default: 'notBool',
-            },
-          ],
-        },
-      });
-
-      if (badDefaultDa.status >= 400) {
-        console.error(
-          'POST /das failed as expected:',
-          badDefaultDa.status,
-          badDefaultDa.json ?? badDefaultDa.text,
-        );
-      }
-      expect(badDefaultDa.status).toBe(400);
-
-      // Create DA with bad params Enum (non string or number values)
-      const badEnumDa = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas/${fdaId}/das`,
-        headers: { 'Fiware-Service': service },
-        body: {
-          id: `${daId2}_badType`,
-          description: 'get user',
-          query: daQuery,
-          params: [
-            {
-              name: 'name',
-              type: 'Text',
-              default: 'carlos',
-              enum: [true, false],
-            },
-          ],
-        },
-      });
-
-      if (badEnumDa.status >= 400) {
-        console.error(
-          'POST /das failed as expected:',
-          badEnumDa.status,
-          badEnumDa.json ?? badEnumDa.text,
-        );
-      }
-      expect(badEnumDa.status).toBe(400);
-
-      // Create DA with bad params range
-      const createBadDa = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas/${fdaId}/das`,
-        headers: { 'Fiware-Service': service },
-        body: {
-          id: `${daId2}_badRange`,
-          description: 'get user',
-          query: daQuery,
-          params: [
-            {
-              name: 'minAge',
-              type: 'Number',
-              default: 25,
-              range: ['badRange', 50],
-            },
-          ],
-        },
-      });
-
-      if (createBadDa.status >= 400) {
-        console.error(
-          'POST /das failed as expected:',
-          createBadDa.status,
-          createBadDa.json ?? createBadDa.text,
-        );
-      }
-      expect(createBadDa.status).toBe(400);
-
-      // Create DA with bad params range (bad order)
-      const createBadRangeDa = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas/${fdaId}/das`,
-        headers: { 'Fiware-Service': service },
-        body: {
-          id: `${daId2}_badRangeOrder`,
-          description: 'get user',
-          query: daQuery,
-          params: [
-            {
-              name: 'minAge',
-              type: 'Number',
-              default: 25,
-              range: [60, 50],
-            },
-          ],
-        },
-      });
-
-      if (createBadRangeDa.status >= 400) {
-        console.error(
-          'POST /das failed as expected:',
-          createBadRangeDa.status,
-          createBadRangeDa.json ?? createBadRangeDa.text,
-        );
-      }
-      expect(createBadRangeDa.status).toBe(400);
-
-      // Create DA with bad params range (bad length)
-      const badRangeLengthDa = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas/${fdaId}/das`,
-        headers: { 'Fiware-Service': service },
-        body: {
-          id: `${daId2}_badRangeLength`,
-          description: 'get user',
-          query: daQuery,
-          params: [
-            {
-              name: 'minAge',
-              type: 'Number',
-              default: 25,
-              range: [40, 50, 60],
-            },
-          ],
-        },
-      });
-
-      if (badRangeLengthDa.status >= 400) {
-        console.error(
-          'POST /das failed as expected:',
-          badRangeLengthDa.status,
-          badRangeLengthDa.json ?? badRangeLengthDa.text,
-        );
-      }
-      expect(badRangeLengthDa.status).toBe(400);
-
-      // Query with name outside of the enum
-      const enumQueryRes = await httpReq({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaId, daId2, {
-          minAge: 25,
-          name: 'fakeNAme',
-        }),
-        headers: { 'Fiware-Service': service },
-      });
-
-      if (enumQueryRes.status >= 400) {
-        console.error(
-          'GET /{visibility}/fdas/{fdaId}/das/{daId}/data failed as expected:',
-          enumQueryRes.status,
-          enumQueryRes.json ?? enumQueryRes.text,
-        );
-      }
-      expect(enumQueryRes.status).toBe(400);
-
-      // Query with age outside of the range
-      const rangeQueryRes = await httpReq({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaId, daId2, {
-          minAge: 60,
-          name: 'ana',
-        }),
-        headers: { 'Fiware-Service': service },
-      });
-
-      if (rangeQueryRes.status >= 400) {
-        console.error(
-          'GET /{visibility}/fdas/{fdaId}/das/{daId}/data failed as expected:',
-          rangeQueryRes.status,
-          rangeQueryRes.json ?? rangeQueryRes.text,
-        );
-      }
-      expect(rangeQueryRes.status).toBe(400);
-
-      // Good query using default values
-      const defaultsQueryRes = await httpReq({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaId, daId2, {
-          name: 'carlos',
-        }),
-        headers: { 'Fiware-Service': service },
-      });
-
-      if (defaultsQueryRes.status >= 400) {
-        console.error(
-          'GET /{visibility}/fdas/{fdaId}/das/{daId}/data failed:',
-          defaultsQueryRes.status,
-          defaultsQueryRes.json ?? defaultsQueryRes.text,
-        );
-      }
-      expect(defaultsQueryRes.status).toBe(200);
-      expect(defaultsQueryRes.json).toEqual([
-        { id: '3', name: 'carlos', age: '40' },
-      ]);
-
-      // Query without required value
-      const requiredQueryRes = await httpReq({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaId, daId2, {
-          minAge: 25,
-        }),
-        headers: { 'Fiware-Service': service },
-      });
-
-      if (requiredQueryRes.status >= 400) {
-        console.error(
-          'GET /{visibility}/fdas/{fdaId}/das/{daId}/data failed as expected:',
-          requiredQueryRes.status,
-          requiredQueryRes.json ?? requiredQueryRes.text,
-        );
-      }
-      expect(requiredQueryRes.status).toBe(400);
-
-      // Query without proper type param
-      const typeQueryRes = await httpReq({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaId, daId2, {
-          minAge: 'text',
-          name: 'carlos',
-        }),
-        headers: { 'Fiware-Service': service },
-      });
-
-      if (typeQueryRes.status >= 400) {
-        console.error(
-          'GET /{visibility}/fdas/{fdaId}/das/{daId}/data failed as expected:',
-          typeQueryRes.status,
-          typeQueryRes.json ?? typeQueryRes.text,
-        );
-      }
-      expect(typeQueryRes.status).toBe(400);
-
-      // Query without proper date (ISO8601)
-      const dateQueryRes = await httpReq({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaId, daId2, {
-          name: 'carlos',
-          timeinstant: '2020-08-17 18:25:28.332+01:00',
-        }),
-        headers: { 'Fiware-Service': service },
-      });
-
-      if (dateQueryRes.status >= 400) {
-        console.error(
-          'GET /{visibility}/fdas/{fdaId}/das/{daId}/data failed as expected:',
-          dateQueryRes.status,
-          dateQueryRes.json ?? dateQueryRes.text,
-        );
-      }
-      expect(dateQueryRes.status).toBe(400);
-
-      // Query without proper date (ISO8601)
-      const boolQueryRes = await httpReq({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaId, daId2, {
-          name: 'carlos',
-          authorized: 'notBool',
-        }),
-        headers: { 'Fiware-Service': service },
-      });
-
-      if (boolQueryRes.status >= 400) {
-        console.error(
-          'GET /{visibility}/fdas/{fdaId}/das/{daId}/data failed as expected:',
-          boolQueryRes.status,
-          boolQueryRes.json ?? boolQueryRes.text,
-        );
-      }
-      expect(boolQueryRes.status).toBe(400);
-
-      // Query with proper type coercion: Number from string
-      const numberCoercionRes = await httpReq({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaId, daId2, {
-          minAge: 30,
-          name: 'carlos',
-        }),
-        headers: { 'Fiware-Service': service },
-      });
-
-      expect(numberCoercionRes.status).toBe(200);
-      expect(numberCoercionRes.json).toEqual([
-        { id: '3', name: 'carlos', age: '40' },
-      ]);
-
-      // Query with proper type coercion: Boolean from string
-      const booleanCoercionRes = await httpReq({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaId, daId2, {
-          name: 'ana',
-          authorized: true,
-        }),
-        headers: { 'Fiware-Service': service },
-      });
-
-      expect(booleanCoercionRes.status).toBe(200);
-      expect(booleanCoercionRes.json).toEqual([
-        { id: '1', name: 'ana', age: '30' },
-      ]);
-
-      // Query with proper type coercion: DateTime from string (ISO8601)
-      const dateTimeCoercionRes = await httpReq({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaId, daId2, {
-          name: 'carlos',
-          timeinstant: '2020-08-17T18:25:28.332Z',
-        }),
-        headers: { 'Fiware-Service': service },
-      });
-
-      expect(dateTimeCoercionRes.status).toBe(200);
-      expect(dateTimeCoercionRes.json).toEqual([
-        { id: '3', name: 'carlos', age: '40' },
-      ]);
-    });
-
-    test('GET /{visibility}/fdas/{fdaId}/das/{daId}/data returns NDJSON when Accept: application/x-ndjson', async () => {
-      const queryRes = await httpReq({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaId, daId, {
-          minAge: 25,
-        }),
-        headers: {
-          'Fiware-Service': service,
-          Accept: 'application/x-ndjson',
-        },
-      });
-
-      if (queryRes.status >= 400) {
-        console.error(
-          'GET /{visibility}/fdas/{fdaId}/das/{daId}/data NDJSON failed:',
-          queryRes.status,
-          queryRes.text,
-        );
-      }
-      expect(queryRes.status).toBe(200);
-      expect(queryRes.text).toBeDefined();
-
-      // Parse NDJSON: split by newline and parse each line as JSON
-      const lines = queryRes.text.split('\n').filter((line) => line.trim());
-      expect(lines.length).toBe(2);
-
-      const row1 = JSON.parse(lines[0]);
-      const row2 = JSON.parse(lines[1]);
-
-      expect(row1).toEqual({ id: 1, name: 'ana', age: 30 });
-      expect(row2).toEqual({ id: 3, name: 'carlos', age: 40 });
-    });
-
-    test('GET /{visibility}/fdas/{fdaId}/das/{daId}/data returns CSV when Accept: text/csv', async () => {
-      const fdaCsvId = 'fda_accept_csv';
-      const daCsvId = 'da_accept_csv';
-
-      const createFda = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas`,
-        headers: {
-          'Fiware-Service': service,
-          'Fiware-ServicePath': servicePath,
-        },
-        body: {
-          id: fdaCsvId,
-          query: 'SELECT id, name, age FROM public.users ORDER BY id',
-          description: 'accept csv fixture',
-        },
-      });
-
-      expect(createFda.status).toBe(202);
-      await waitUntilFDACompleted({ baseUrl, service, fdaId: fdaCsvId });
-
-      const createDa = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas/${fdaCsvId}/das`,
-        headers: { 'Fiware-Service': service },
-        body: {
-          id: daCsvId,
-          description: 'accept csv da fixture',
-          query: `
-            SELECT id, name, age
-            WHERE age > $minAge
-            ORDER BY id
-          `,
-          params: [{ name: 'minAge', type: 'Number', required: true }],
-        },
-      });
-
-      expect(createDa.status).toBe(201);
-
-      const res = await httpReqRaw({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaCsvId, daCsvId, {
-          minAge: 25,
-        }),
-        headers: {
-          'Fiware-Service': service,
-          Accept: 'text/csv',
-        },
-      });
-
-      if (res.status >= 400) {
-        console.error(
-          'GET /{visibility}/fdas/{fdaId}/das/{daId}/data Accept=text/csv failed:',
-          res.status,
-          res.text,
-        );
-      }
-
-      expect(res.status).toBe(200);
-      expect(res.headers['content-type']).toContain('text/csv');
-      expect(res.headers['content-disposition']).toContain('results.csv');
-
-      const lines = res.text
-        .split('\n')
-        .map((line) => line.trim())
-        .filter(Boolean);
-      expect(lines[0]).toBe('id,name,age');
-      expect(lines[1]).toBe('1,ana,30');
-      expect(lines[2]).toBe('3,carlos,40');
-    });
-
-    test('GET /{visibility}/fdas/{fdaId}/das/{daId}/data returns XLSX when Accept requests spreadsheet mime', async () => {
-      const fdaXlsId = 'fda_accept_xls';
-      const daXlsId = 'da_accept_xls';
-
-      const createFda = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas`,
-        headers: {
-          'Fiware-Service': service,
-          'Fiware-ServicePath': servicePath,
-        },
-        body: {
-          id: fdaXlsId,
-          query: 'SELECT id, name, age FROM public.users ORDER BY id',
-          description: 'accept xls fixture',
-        },
-      });
-
-      expect(createFda.status).toBe(202);
-      await waitUntilFDACompleted({ baseUrl, service, fdaId: fdaXlsId });
-
-      const createDa = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas/${fdaXlsId}/das`,
-        headers: { 'Fiware-Service': service },
-        body: {
-          id: daXlsId,
-          description: 'accept xls da fixture',
-          query: `
-            SELECT id, name, age
-            WHERE age > $minAge
-            ORDER BY id
-          `,
-          params: [{ name: 'minAge', type: 'Number', required: true }],
-        },
-      });
-
-      expect(createDa.status).toBe(201);
-
-      const res = await httpReqRaw({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaXlsId, daXlsId, {
-          minAge: 25,
-        }),
-        headers: {
-          'Fiware-Service': service,
-          Accept:
-            'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
-        },
-      });
-
-      if (res.status >= 400) {
-        console.error(
-          'GET /{visibility}/fdas/{fdaId}/das/{daId}/data Accept=xlsx failed:',
-          res.status,
-          res.text,
-        );
-      }
-
-      expect(res.status).toBe(200);
-      expect(res.headers['content-type']).toContain(
-        'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
-      );
-      expect(res.headers['content-disposition']).toContain('results.xlsx');
-
-      // XLSX is a ZIP-based format and starts with PK magic bytes.
-      expect(res.buffer[0]).toBe(0x50);
-      expect(res.buffer[1]).toBe(0x4b);
-      expect(res.buffer.length).toBeGreaterThan(100);
-    });
-
-    test('GET /{visibility}/fdas/{fdaId}/das/{daId}/data rejects outputType query param', async () => {
-      const fdaJsonId = 'fda_accept_json';
-      const daJsonId = 'da_accept_json';
-
-      const createFda = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas`,
-        headers: {
-          'Fiware-Service': service,
-          'Fiware-ServicePath': servicePath,
-        },
-        body: {
-          id: fdaJsonId,
-          query: 'SELECT id, name, age FROM public.users ORDER BY id',
-          description: 'accept json fixture',
-        },
-      });
-
-      expect(createFda.status).toBe(202);
-      await waitUntilFDACompleted({ baseUrl, service, fdaId: fdaJsonId });
-
-      const createDa = await httpReq({
-        method: 'POST',
-        url: `${baseUrl}/${visibility}/fdas/${fdaJsonId}/das`,
-        headers: { 'Fiware-Service': service },
-        body: {
-          id: daJsonId,
-          description: 'accept json da fixture',
-          query: `
-            SELECT id, name, age
-            WHERE age > $minAge
-            ORDER BY id
-          `,
-          params: [{ name: 'minAge', type: 'Number', required: true }],
-        },
-      });
-
-      expect(createDa.status).toBe(201);
-
-      const res = await httpReq({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaJsonId, daJsonId, {
-          minAge: 25,
-          outputType: 'html',
-        }),
-        headers: { 'Fiware-Service': service },
-      });
-
-      expect(res.status).toBe(400);
-      expect(res.json.error).toBe('BadRequest');
-      expect(res.json.description).toBe(
-        'Invalid fields in request query, check your request',
-      );
-    });
-
-    test('GET /{visibility}/fdas/{fdaId}/das/{daId}/data returns 406 for unsupported Accept header', async () => {
-      const res = await httpReq({
-        method: 'GET',
-        url: buildDaDataUrl(baseUrl, servicePath, fdaId, daId, {
-          minAge: 25,
-        }),
-        headers: {
-          'Fiware-Service': service,
-          Accept: 'video/mpg4',
-        },
-      });
-
-      expect(res.status).toBe(406);
-      expect(res.json.error).toBe('NotAcceptable');
-      expect(res.json.description).toContain(
-        'Accept header must allow application/json',
-      );
     });
 
     test('GET /{visibility}/... returns 400 for an invalid visibility value', async () => {
